app/models/user.py
from sqlalchemy import Column, Integer, String
from flask_sqlalchemy import SQLAlchemy

from .. import db # Import the db instance from app/__init__.py
from datetime import datetime

class User(db.Model):
    __tablename__ = 'users'

    id = db.Column(db.Integer, primary_key=True)
    email = db.Column(db.String(120), unique=True, nullable=False)
    password_hash = db.Column(db.String(256), nullable=False) # Increased length for modern hashes
    date_created = db.Column(db.DateTime, default=datetime.utcnow)

    def __repr__(self):
        return f"<User {self.email}>"