requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
python-dotenv
pytest
pytest-flask
python-dotenv